refactor(server): 收编散装 ProjectManager 构造为统一懒加载单例
lib/project_manager.py 新增 get_project_manager()，全库约 17 处散装的
ProjectManager(app_data_dir()) 构造（13 个 router、generation_tasks、
config resolver、server/app 启动流程）统一收编为经此函数取用，保留既有
late-binding + monkeypatch 测试约定；cost_estimation.py 与 assets.py 原先
直接 monkeypatch 模块级 pm 变量的测试一并收敛到同一约定。
