+++ tests/server/test_reference_videos_router_ad.py
@@ -88,7 +88,6 @@ def ad_client(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> TestClient:
     from server.routers import reference_videos as router_mod
 
     custom_pm = ProjectManager(projects_root)
-    monkeypatch.setattr(router_mod, "pm", custom_pm)
     monkeypatch.setattr(router_mod, "get_project_manager", lambda: custom_pm)
 
     # 供应商时长上限解析与队列都打桩：路由测试只看入参与持久化结果
