+++ tests/test_cost_estimation_router.py
@@ -17,7 +17,7 @@ def _make_app():
 
 
 def _mock_pm(**overrides):
-    """Create a mock to replace the module-level ``pm`` singleton."""
+    """Create a mock to replace the ``get_project_manager()`` singleton getter."""
     mock = MagicMock()
     for k, v in overrides.items():
         setattr(mock, k, MagicMock(return_value=v))
@@ -26,7 +26,7 @@ def _mock_pm(**overrides):
 
 class TestCostEstimationRouter:
     def test_project_not_found_returns_404(self):
-        with patch.object(cost_estimation, "pm", _mock_pm(project_exists=False)):
+        with patch.object(cost_estimation, "get_project_manager", lambda: _mock_pm(project_exists=False)):
             with TestClient(_make_app()) as client:
                 resp = client.get("/api/v1/projects/nonexistent/cost-estimate")
         assert resp.status_code == 404
@@ -46,7 +46,7 @@ class TestCostEstimationRouter:
         mock_pm = _mock_pm(project_exists=True, load_project={"episodes": []})
 
         with (
-            patch.object(cost_estimation, "pm", mock_pm),
+            patch.object(cost_estimation, "get_project_manager", lambda: mock_pm),
             patch.object(cost_estimation, "CostEstimationService") as MockService,
         ):
             MockService.return_value.compute = AsyncMock(return_value=fake_result)
